Refactor: extract on new files github and changelog
